* Use the same dto for user and family profile. * Mark some columns as nullable.
