+++ backend/src/StamAcasa.Common/Models/RelationshipTypes.cs
@@ -1,0 +1,11 @@
+﻿namespace StamAcasa.Common.Models
+{
+    public enum RelationshipTypes
+    {
+        Parent,
+        Child,
+        Grandparent,
+        Neighbor,
+        Other
+    }
+}
