+++ backend/src/StamAcasa.Common/Models/User.cs
@@ -14,33 +14,23 @@ namespace StamAcasa.Common.Models {
         public int Id { get; set; }
         public int? ParentId { get; set; }
         public string Sub { get; set; }
+        public string Email { get; set; }
         public string Name { get; set; }
         public string Surname { get; set; }
         public string PhoneNumber { get; set; }
         public int Age { get; set; }
         public Gender Gender { get; set; }
-        public string PreexistingMedicalCondition { get; set; }
-        public QuarantineStatus QuarantineStatus { get; set; }
         public string County { get; set; }
         public string City { get; set; }
-        public string Country { get; set; }
-        public string Email { get; set; }
-        public string Disability { get; set; }
-        public RelationshipTypes RelationshipType { get; set; }
+        public string PreexistingMedicalCondition { get; set; }
+        public QuarantineStatus QuarantineStatus { get; set; }
         public bool Smoker { get; set; }
         public bool LivesWithOthers { get; set; }
-        public QuarantineStatus QuarantineStatusOthers { get; set; }
+        public QuarantineStatus? QuarantineStatusOthers { get; set; }
+        public RelationshipTypes? RelationshipType { get; set; }
 
         public virtual User ParentUser { get; set; }
         public virtual HashSet<User> DependentUsers { get; set; }
         public virtual HashSet<Form> Forms { get; set; }
-    }
-
-    public enum RelationshipTypes
-    {
-        Parent,
-        Child,
-        Grandparent,
-        Other
     }
 }
