+++ backend/src/StamAcasa.Common/Services/IUserService.cs
@@ -8,7 +8,7 @@ namespace StamAcasa.Common.Services {
     public interface IUserService
     {
         Task<UserInfo> AddOrUpdateUserInfo(UserProfileDTO user);
-        Task<UserInfo> AddOrUpdateDependentInfo(FamilyProfileDTO user, string parentSub);
+        Task<UserInfo> AddOrUpdateDependentInfo(UserProfileDTO user, string parentSub);
         Task<UserInfo> GetUserInfo(string sub);
         Task<UserInfo> GetUserInfo(int id);
         Task<IEnumerable<UserInfo>> GetDependentInfo(string sub);
